Added trivial generator to example language
This works around the fact that the EVM scheduler can be called before
the Xtext builder terminates, thus calling the EVM rules in an
inconsistent state. This workaround ensures that after the builder is
executed, the EVM scheduler is triggered again, making the rule state
eventually consistent with the Xtext Index.